+++ 03_scan/kernel.cu
@@ -3,15 +3,15 @@
 #include "../utils/tracer.h"
 #include <vector>
 
-__global__ void scan_kernel(const float* a, float* c, int size) {
+__global__ void scan_kernel(const float* a, float* c, long long size) {
     // TODO: Implement exclusive prefix sum
 }
 
-std::vector<LaunchConfig> launch_scan(const float* a, float* c, int size) {
+std::vector<LaunchConfig> launch_scan(const float* a, float* c, long long size) {
     global_tracer.trace("Entering launch_scan (Student)");
     
     int threadsPerBlock = 256;
-    int blocksPerGrid = (size + threadsPerBlock - 1) / threadsPerBlock;
+    long long blocksPerGrid = (size + threadsPerBlock - 1) / threadsPerBlock;
     
     global_tracer.trace("Launching scan_kernel (Student)");
     // scan_kernel<<<blocksPerGrid, threadsPerBlock>>>(a, c, size);
