+++ 03_scan/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_scan(const float* a, float* c, int size);
+std::vector<LaunchConfig> launch_scan(const float* a, float* c, long long size);
