+++ 05_stencil/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_stencil(const float* a, float* c, int size, int radius);
+std::vector<LaunchConfig> launch_stencil(const float* a, float* c, long long size, int radius);
