+++ 06_scatter_gather/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_gather(const float* source, float* dest, const int* indices, int size);
+std::vector<LaunchConfig> launch_gather(const float* source, float* dest, const int* indices, long long size);
