+++ 07_gemm/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_gemm(const float* a, const float* b, float* c, int n);
+std::vector<LaunchConfig> launch_gemm(const float* a, const float* b, float* c, long long n);
