+++ 08_stream_compaction/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_compaction(const int* a, int* c, int* count, int size);
+std::vector<LaunchConfig> launch_compaction(const int* a, int* c, int* count, long long size);
