+++ 09_radix_sort/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_radix_sort(const int* a, int* c, int size);
+std::vector<LaunchConfig> launch_radix_sort(const int* a, int* c, long long size);
