+++ 12_segmented_scan/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_segmented_scan(const float* a, const int* flags, float* c, int size);
+std::vector<LaunchConfig> launch_segmented_scan(const float* a, const int* flags, float* c, long long size);
