+++ 15_conv2d/reference_kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_reference_conv2d(const float* a, const float* filter, float* c, int width, int height);
+std::vector<LaunchConfig> launch_reference_conv2d(const float* a, const float* filter, float* c, long long width, long long height);
